+++ package.json
@@ -40,6 +40,7 @@
     "react-hook-form": "^7.43.2",
     "react-icons": "^4.8.0",
     "recoil": "^0.7.7",
+    "recoilize": "^3.1.6",
     "swr": "^2.0.4",
     "typescript": "4.9.5",
     "use-immer": "^0.8.1",
