+++ src/page/edit/draw/HandRange/HandOne/HandOne.stories.tsx
