+++ src/page/edit/draw/HandRange/HandSquare/HandSquare.stories.tsx
