+++ src/page/edit/draw/HandRange/HandSquare/index.tsx
